Use search param within selected input context
