+++ front/app/containers/Admin/projects/project/analysis/SelectedInputContext/index.tsx
@@ -1,4 +1,12 @@
-import React, { Dispatch, createContext, useContext, useState } from 'react';
+import React, {
+  Dispatch,
+  createContext,
+  useContext,
+  useState,
+  useEffect,
+} from 'react';
+import { useSearchParams } from 'react-router-dom';
+import { removeSearchParams } from 'utils/cl-router/removeSearchParams';
 
 const Context = createContext<{
   selectedInputId: string | null;
@@ -10,6 +18,20 @@ const Context = createContext<{
 
 const SelectedInputContext = ({ children }) => {
   const [selectedInputId, setSelectedInputId] = useState<string | null>(null);
+  const [search] = useSearchParams();
+  const inputId = search.get('selected-input-id');
+
+  useEffect(() => {
+    if (inputId) {
+      setSelectedInputId(inputId);
+      const element = document.getElementById(`input-${inputId}`);
+      if (element) {
+        element.scrollIntoView({ behavior: 'smooth' });
+      }
+    }
+    removeSearchParams(['selected-input-id']);
+  }, [inputId]);
+
   const contextValue = { selectedInputId, setSelectedInputId };
   return <Context.Provider value={contextValue}>{children}</Context.Provider>;
 };
